fixes(paths): The paths were too unique...
I had to normalize the paths due to the fact that sometimes a path to a
file would be absolute or relative to the project depending on how you
opened up the file.  To prevent this from creating duplicates in a mark
list, I normalized the path before doing anything.