Add further logging, fix callbacks
The callbacks are in a nested table, so they weren't actually called. If
there aren't more events then callbacks for "changed" don't need to be
nested in the callbacks table. Alternatively emitChanged() could take an
event name: emit("changed") and run callbacks for a specific event or
take a table of events and run for all, emit({"changed", "navigate"}).