fix(ui): Sets buffer as 'listed' when navigating
According to the documentation for `vim.fn.bufnr` (`:h bufnr`), the
created buffer will have the option 'unlisted' by default.

This means that when navigating to a file, if that file wasn't
previously open, the new buffer won't appear in buffers list.

This fix set the buffer to be 'listed' when navigating.