fix(marks): fix missing entry-index
This commit fixes the missing `entry.index` - in the marks-extension for
telescope - which caused the index to always show `nil` - or rather
`n…` - due to a fixed width of 2, in the dialogue.

This commit also added a dynamic width and left-padding for the
index-column, so the index is being right-aligned.