feat(picker): sort by index
This change allows the telescope-picker to sort by index. If one of the
search-terms equals an index, it'll get a better score. Partial matches
of an index, will get a slightly better score. Matches with the main
entry - i.e. the filename - will always have the best score; showing
them before an index-match.